requirements.txt
amqp==5.1.1
async-timeout==4.0.3
backports.zoneinfo==0.2.1
bidict==0.22.1
billiard==4.1.0
blinker==1.6.2
celery==5.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.4.2
eventlet==0.33.3
Flask==2.3.3
Flask-SocketIO==5.3.5
greenlet==2.0.2
gunicorn==21.2.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.1
MarkupSafe==2.1.3
numpy==1.21.4
packaging==23.1
prompt-toolkit==3.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.6.1
python-socketio==5.8.0
redis==5.0.0
six==1.16.0
typing_extensions==4.7.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.7
zipp==3.16.2
